Fix compile against qt6.4
"error C2397: conversion from 'qsizetype' to 'int' requires a narrowing
conversion"

+++ src/kitemviews/private/kdirectorycontentscounterworker.cpp
@@ -103,7 +103,7 @@ KDirectoryContentsCounterWorker::CountResult KDirectoryContentsCounterWorker::su
     } else {
         filters |= QDir::AllEntries;
     }
-    return {dir.entryList(filters).count(), 0};
+    return {static_cast<int>(dir.entryList(filters).count()), 0};
 #else
 
     const uint maxRecursiveLevel = DetailsModeSettings::directorySizeCount() ? 1 : DetailsModeSettings::recursiveDirectorySizeLimit();
